ci: Script for checking-out experimental types (#507)
# Description

Added a script for creating a new feature branch with types generated
from unmerged A2A spec changes.

We need to begin development on several upcoming spec changes that are
currently unmerged but are highly likely to be included in the next
release, such as the task/list method. To do that, we need updated
types, which this script provides.

# Prerequisites
- [x] Follow the [`CONTRIBUTING`
Guide](https://github.com/a2aproject/a2a-python/blob/main/CONTRIBUTING.md).
- [x] Make your Pull Request title in the
<https://www.conventionalcommits.org/> specification.
- Important Prefixes for
[release-please](https://github.com/googleapis/release-please):
- `fix:` which represents bug fixes, and correlates to a
[SemVer](https://semver.org/) patch.
- `feat:` represents a new feature, and correlates to a SemVer minor.
- `feat!:`, or `fix!:`, `refactor!:`, etc., which represent a breaking
change (indicated by the `!`) and will result in a SemVer major.
- [x] Ensure the tests and linter pass (Run `bash scripts/format.sh`
from the repository root to format)
- [x] Appropriate docs were updated (if necessary)

---------

Co-authored-by: lkawka <lkawka@google.com>